Import Commit: mass import fll project
This is not an import from the svn repository, but instead an import from the latest individual 0.2.* release for each project.
This also has a few patches that I had applied or backported from kevux to fix bugs in the latest official releases.

I did this because I am completely out of sync with whatever I was doing in the project.
I was probably in the middle of making some major changes, but I do not remember what they were.
Therefore, I am starting on a hopefully stable base to pick up from.