Update: implement fll_execute_program and fix fll_execute_path
This adds fll_execute_program and changes how fll_execute_path works to function properly.
Prior to this change, fll_execute_path did not set the first argument which would cause the programs argv[0] to not be equivalent to the program name.

This also corrects a few mistakes where level_1 syntax and comments were not changed to level_2.