Update: add pipe console parameter
The fll standard console arguments now inclide +p, ++pipe.
This pipe command requires a single numeric argument that represents the size of the pipe.
When this is specified, it tells the application to read from pipe in byte chunks of the specified size.
The default pipe size is left up to the application when this is not specified.