Update: rewrite basic list object and content read functions
Follow the new approach.
Also, it occured to me that there is no reason to support quoted object names for a basic list.

The reasoning behind this is that the basic list object always terminates on a newline and there cannot be any content on that newline.
If there is no content, then there is no reason to protect the object by wrapping it in quotes as done in basic an extended.
Quote support has been removed and as a result the implementation of the object and content reads are much simpler.