Bugfix: ipv6 icmp is different than ipv4 icmp, adjust rules accordingly
The ipv6 in iptables uses a separate command called icmpv6, with the parameter called icmpv6-type.
Adjust the default settings files to use ipv4 icmp only with iptables and ipv6 icmp only with ip6tables.

This also updates the icmp types for ipv6, adding comments on what each type is.