Bugfix: programs incorrectly reporting version
I noticed, unfortunately after the release and not before, that the proper version number was not being reported.
This seems to have been an oversight.

To reduce the chances of this happening, I decided to make the next version number be the default.
I will make a release after a month or two fixing this.