Update: add signal code support to f_errors and move error code starting point to 166
Signals are used as the high order bits in a process that exits, which is where I provide error and similar return codes.
For this reason I am adding wrappers to the x86 signal codes in f_errors instead of a separate custom project such as the non-existant f_signals.

It looks like bash mixes the exit codes and signals into the numbers 129 to 165, so start with 166 as well.