Bugfix: revert compiler workaround 5bdaa0b, place arguments after source
The problem turns out to be that, for gcc at the very least, linker options such as -lc are not properly passed to the linker when it is before the source files.
Using the examples:
  1) gcc -lc main.c
  2) gcc main.c -lc

In case of #1, the -lc is not passed to the linker.
In case of #2, the -lc is passed to the linker.

All arguments are now being moved after the source parts.

see: 5bdaa0b Workaround: compilers on some systems seem to be sensitive to the order of arguments
see: http://gcc.gnu.org/onlinedocs/gcc/Link-Options.html