Bugfix: fix order of library linkage for firewall build settings
The ar command is sensitive to the order in which libraries are added.
Prior to this change, building a static library could have dependencies used before they are defined due to the order in which libraries were added.
Change the order so that libraries are to the right of their dependencies in the command line argument order