Bugfix: reverse order of library links
It seems I got the order backwards.
I was using -lc and all the lowest level libraries on the left of the libraries that depend on it.
This caused unobvious linkage errors down the road.
Reversing the order prevented the problem and allowed for static libraries to link.

It is possible the order issue is with ar and not with gcc.