Progress: miscellaneous changes
I was working on stuff a while back and was taken away from the project for quite a while.
These are some uncommitted changes that I would rather commit immediately and then move forward.

There are some other changes I will follow up in separate commits, such as the socket work and the init work.