Feature: add install.sh helper script
Provide a very basic install helper script for systems (or users) that do not have or do not want to use their package system to install the compiled files.

This is only intended to help someone install.
It is still recommended that a distribution or packager should povide their own tools to perform the installation.