Cleanup: fix comments and remove commented out memset
According to the standards that I have read, calloc() guarantees zero'd memory, so memset is not needed.
It so happens that the memset was commented out already, so just remove it.

There are some memset related comments that incorrectly use the word 'bool' when the word 'char' should be used.