Update: add basic featureless settings specifications documentation
This was originally defined online, but I had lost the data at some point.
The internet archive didn't cache the page and the data lost happened during the time I was changing the website.

The specifications were simple enough to be rewritten, but I have done so in haste.
Some of the specifications will be documented at a later date.

I do not remember what advanced reverse mapping was intended to be, so I replaced it with the custom json-like specification.
The goal of the json-like specification is to write it such that it is still valid json (minus any comments).
It is currently added to preserve the idea but there are no immediate plans to implement this specification.