Revert: remove custom type defines, including use of f_autochar
The intent of these custom types is to allow for easily replacing what the type represents.
I have decided that this extent of doing such is overkill.
Revert the custom types to instead rely on the C/C++ defined types, where appropriate.

After having gained more experience with utf-8 an other unicode standards, I have come to conclude that the wchar type should not be used.
This project will support the char type instead of the wchar type.
All uses have been replaced with char.