Cleanup: Use new comment structure at the tops of files
Just put the comments in the header files.
Simplify the comment stucture.
The git repository provides the authors/credits, no need to explicitly do so in the files anymore.
Do not provide the micro version, instead provide the major and minor versions in the files (which effectively represent the API version).