Update: add build_libraries_fll build setting
Add a new parameter to make it easier to switch between individual compilation, level compilation, and monolithic compilation.
This parameter will only be used for fll-specific libraries.

For programs (aka: "level 3") commented out examples for linking against level-based and monolithic are provided.
The configurations can then be easily swapped out by a couple of simple sed statements, such as:
  - sed -i -e 's|^build_libraries_fll\>|#&|g' data/build/settings
  - sed -i -e 's|^#build_libraries_fll-level\>|build_libraries_fll|g' data/build/settings

The above example will switch to the level based compiling, while disabling the individual compiling.

The level based compiling compiles each of the levels 0, 1, and 2 as a single library for each level, resulting in libraries such as:
  - libfll_0-0.5.0.so
  - libfll_1-0.5.0.so
  - libfll_2-0.5.0.so

The monolithic based compiling compiles all of the levels 0, 1, and 2 as a single library, resulting in libraries such as:
  - libfll-0.5.0.so

The standard names of individual, level, and monolithic do not overlap by default and can therefoe be installed side-by-side.