Feature: add 'dependency' generation to package.sh and update all dependencies
This will help ensure that dependencies will be accurate and less error prone (so long as the script is run after making changes to any projects).

Dependencies are processed from the projects individual data/build/dependencies file.
The order of the dependencies does matter and is processed from top to bottom.
The 4 core dependencies must be first if they are depended on (and in this order: f_type, f_status, f_memory, and f_string).

The dependency generation for individual projects will generate the libraries to link against if a given project has any library source files.
Linking is done in highest level to lowest level to help ensure no linking errors happen.

The dependency generation for level projects and monolithic project are done based on the library and header sources.
All files specified in the build_sources_library and build_sources_headers are generated.

This script has been run and the dependency updates generated by this script are included in this commit.