Regression: display "+" and "++" and not "-" and "--" for special parameter options
When I wrote fll_program_print_help_option() I completely forgot to provide a way to set eithe "-" or "+" and "--" or "++".
This resulted in the "--help" display of the options to incorrectly print using "-" and "--".

Add additional function parameters to allow setting the symbols when calling fll_program_print_help_option().