Update: redesign byte_dump --last to be inclusive
The --last parameter should be inclusive for consistency with the rest of the project, namely FSS.

By redesigning the --last to internally be represented as a relative offset, the behavior can be simplified.
Doing this also adds support for potential future designs where a --length parameter may be provided that is a relative size parameter instead of an absolute one like --last.