Progress: begin re-implementing fake program (Featureless Make)
The goal here will be to mimic the current FSS build scripts fist.
Then update the FSS build scripts to the more generalized design.
Then implement the traditional fakefile process (a makefile alternative).

Currently has "skeleton" mode, which may or may not be implemented.
This was added as a stub just in case I decide to implement it.