Revert "Refactor: f_print_string_dynamic to f_print_string_static"
This reverts commit f225f4059d0e082a7bfae29e042a2120ac4bda33.

It occurred to me that this would make the function names more inconsistent.
This could get very confusing so just stick with "dynamic" in all the names.
"dynamic" is the original term and "static" in isolation is a reserved word.

The type names will remain as both f_string_static and f_string_dynamic.
That inconsistency is desirable due to the explicit meaning communicated by their names.